feat: finish the user interface

+++ MagicalBox/AdminPage.xaml.cs
@@ -50,12 +50,16 @@ namespace MagicalBox
 
         private void removeInfo_Click(object sender, RoutedEventArgs e)
         {
-            using AppDbContext dbContext = new AppDbContext();
             User Single_User = dataList.SelectedItem as User;
-            dbContext.Users.Remove(Single_User);
-            //dataList.Items.Remove(Single_User);
-            dbContext.SaveChanges();
-            initList();
+            if (Single_User != null && Single_User is User)
+            {
+                using AppDbContext dbContext = new AppDbContext();
+                dbContext.Users.Remove(Single_User);
+                //dataList.Items.Remove(Single_User);
+                dbContext.SaveChanges();
+                initList();
+            }
+            else MessageBox.Show("请先选定需要删除信息的乘客！");
         }
 
         private void checkDetail_Click(object sender, RoutedEventArgs e)
@@ -71,15 +75,40 @@ namespace MagicalBox
                     "登机牌号：" + Single_User.BoardCard + "\n\n" +
                     "出发地：" + Single_User.Departure + "\n\n" +
                     "目的地：" + Single_User.Destination + "\n\n" +
-                    "手机机型" + Single_User.PhoneType + "\n\n" +
-                    "手机号码" + Single_User.PhoneNumber + "\n\n" +
-                    "备用联系方式" + Single_User.BackupLink + "\n\n" +
-                    "是否已取回手机" + Single_User.Returned + "\n\n" +
-                    "满意度" + Single_User.Satisfaction + "\n\n" +
-                    "评价" + Single_User.Comment + "\n\n"
+                    "手机机型：" + Single_User.PhoneType + "\n\n" +
+                    "手机号码：" + Single_User.PhoneNumber + "\n\n" +
+                    "备用联系方式：" + Single_User.BackupLink + "\n\n" +
+                    "是否已取回手机：" + Single_User.Returned + "\n\n" +
+                    "满意度：" + Single_User.Satisfaction + "\n\n" +
+                    "评价：" + Single_User.Comment + "\n\n"
                     );
             }
             else MessageBox.Show("请先选定需要查看的乘客！");
+        }
+        private void phoneReturned_Click(object sender, RoutedEventArgs e)
+        {
+            User Single_User = dataList.SelectedItem as User;
+            using AppDbContext dbContext = new AppDbContext();
+            foreach (var item in dbContext.Users.Where(e => e.Id == Single_User.Id))
+            {
+                item.Returned = true;
+            }
+            dbContext.SaveChanges();
+            MessageBox.Show("设备归还成功！");
+            initList();
+        }
+
+        private void easySearch_Click(object sender, RoutedEventArgs e)
+        {
+            //for (int i = 0; i < dataList.Items.Count; i++)
+            //{
+            //    if (dataList.Items[i].BoxId.Text.ToString() == sOrderNo)
+            //    {
+            //        this.dataList.Items[i].Selected = true;//选中行
+            //        this.dataList.EnsureVisible(i);//滚动到指定的行位置
+            //        return;
+            //    }
+            //}
         }
     }
 }
